move the function get_outline -> common.py and remove functions not used
