+++ idstools/view/tf.py
@@ -11,7 +11,6 @@ import matplotlib.pyplot as plt
 import matplotlib.text as mtext
 import numpy as np
 from matplotlib.patches import Patch, Polygon
-from scipy.interpolate import splprep, splev
 
 from idstools.compute.tf import TFCompute
 
@@ -170,227 +169,3 @@ class TFView:
         else:
             ax.set_title("tf")
         return tf_legend
-
-
-def sectional_interpolation(x, y, points_per_section=20):
-    """
-    Apply adaptive interpolation to contour sections with smart
-    gradient control for line segments.
-
-    This function uses a two-pass approach:
-    1. First pass: Identify all line segments in the contour
-    2. Second pass: Apply appropriate interpolation method
-
-    This ensures smooth transitions between line segments and curves.
-
-    Parameters
-    ----------
-    x, y : array_like
-        1D arrays of the polygon vertices, ordered along the loop.
-    points_per_section : int, optional
-        Number of points to interpolate in each section (default: 20).
-
-    Returns
-    -------
-    xs, ys : ndarray
-        Interpolated points along the smoothed contour.
-
-    Notes
-    -----
-    Three interpolation strategies based on segment type:
-
-    1. **Line segments** (both P1 and P2 on a line):
-       - Uses linear interpolation: x = (1-t)*x1 + t*x2
-       - Guarantees perfect straightness (zero deviation)
-
-    2. **Curved sections** (both P1 and P2 on curves):
-       - Uses cubic Hermite with Catmull-Rom gradients
-       - g1 = 0.5 * (P2 - P0), g2 = 0.5 * (P3 - P1)
-       - Provides smooth curve interpolation
-
-    3. **Transitions** (one point on line, one on curve):
-       - Uses cubic Hermite with mixed gradients
-       - Line points use line segment gradient
-       - Curve points use Catmull-Rom gradient
-       - Ensures smooth transition between line and curve
-    """
-    x = np.asarray(x, dtype=float)
-    y = np.asarray(y, dtype=float)
-
-    if len(x) < 4:
-        return x, y
-
-    # Ensure the contour is closed
-    if not (x[0] == x[-1] and y[0] == y[-1]):
-        x = np.append(x, x[0])
-        y = np.append(y, y[0])
-
-    # Remove duplicate closing point for processing
-    x_open = x[:-1]
-    y_open = y[:-1]
-    n = len(x_open)
-
-    # FIRST PASS: Identify all line segments
-    # For each point, store whether it's part of a line segment
-    is_on_line = np.zeros(n, dtype=bool)
-    line_gradients = {}  # Store gradient for each point on a line
-
-    for i in range(n):
-        # Check triplet: P(i-1), P(i), P(i+1)
-        im1 = (i - 1) % n
-        ip1 = (i + 1) % n
-
-        if _is_collinear(x_open[im1], y_open[im1], x_open[i], y_open[i], x_open[ip1], y_open[ip1]):
-            is_on_line[i] = True
-            # Store the line segment gradient at this point
-            line_gradients[i] = (x_open[ip1] - x_open[i], y_open[ip1] - y_open[i])
-
-    # SECOND PASS: Interpolate using computed gradients
-    xs_new = []
-    ys_new = []
-
-    for i in range(n):
-        # Get 4 consecutive points with wraparound
-        p0_idx = i
-        p1_idx = (i + 1) % n
-        p2_idx = (i + 2) % n
-        p3_idx = (i + 3) % n
-
-        x0, y0 = x_open[p0_idx], y_open[p0_idx]
-        x1, y1 = x_open[p1_idx], y_open[p1_idx]
-        x2, y2 = x_open[p2_idx], y_open[p2_idx]
-        x3, y3 = x_open[p3_idx], y_open[p3_idx]
-
-        # Add the starting point of this section
-        xs_new.append(x1)
-        ys_new.append(y1)
-
-        # Calculate gradients at P1 and P2 based on line membership
-        if is_on_line[p1_idx]:
-            # P1 is on a line - use the precomputed line gradient
-            g1_x, g1_y = line_gradients[p1_idx]
-        else:
-            # P1 is on a curve - use Catmull-Rom gradient
-            g1_x, g1_y = 0.5 * (x2 - x0), 0.5 * (y2 - y0)
-
-        if is_on_line[p2_idx]:
-            # P2 is on a line - use the precomputed line gradient
-            g2_x, g2_y = line_gradients[p2_idx]
-        else:
-            # P2 is on a curve - use Catmull-Rom gradient
-            g2_x, g2_y = 0.5 * (x3 - x1), 0.5 * (y3 - y1)
-
-        # Check if both P1 and P2 are on a line segment
-        if is_on_line[p1_idx] and is_on_line[p2_idx]:
-            # Use linear interpolation for line segments
-            for j in range(1, points_per_section):
-                t = j / points_per_section
-                x_interp = (1 - t) * x1 + t * x2
-                y_interp = (1 - t) * y1 + t * y2
-                xs_new.append(x_interp)
-                ys_new.append(y_interp)
-        else:
-            # Apply cubic Hermite interpolation for curves or transitions
-            for j in range(1, points_per_section):
-                t = j / points_per_section
-                t2 = t * t
-                t3 = t2 * t
-
-                # Hermite basis functions
-                h1 = 2 * t3 - 3 * t2 + 1
-                h2 = -2 * t3 + 3 * t2
-                h3 = t3 - 2 * t2 + t
-                h4 = t3 - t2
-
-                x_interp = h1 * x1 + h2 * x2 + h3 * g1_x + h4 * g2_x
-                y_interp = h1 * y1 + h2 * y2 + h3 * g1_y + h4 * g2_y
-
-                xs_new.append(x_interp)
-                ys_new.append(y_interp)
-
-    # Close the contour
-    xs_new.append(xs_new[0])
-    ys_new.append(ys_new[0])
-
-    return np.array(xs_new), np.array(ys_new)
-
-
-def _is_collinear(x1, y1, x2, y2, x3, y3, tolerance=1e-6):
-    """
-    Check if three points are collinear (lie on the same straight line).
-
-    Uses the cross product method: if the cross product of vectors
-    (p1->p2) and (p2->p3) is zero, the points are collinear.
-
-    Parameters
-    ----------
-    x1, y1, x2, y2, x3, y3 : float
-        Coordinates of the three points to check.
-    tolerance : float, optional
-        Numerical tolerance for collinearity check (default: 1e-6).
-
-    Returns
-    -------
-    bool
-        True if the three points are collinear within tolerance.
-    """
-    # Calculate cross product of vectors (p1->p2) and (p2->p3)
-    # Cross product = (x2-x1)*(y3-y2) - (y2-y1)*(x3-x2)
-    cross_product = (x2 - x1) * (y3 - y2) - (y2 - y1) * (x3 - x2)
-
-    # Check if cross product is close to zero
-    return abs(cross_product) < tolerance
-
-
-def periodic_closed_spline(x, y, num=64, smooth=0.0, k=2):
-    """
-    Build a periodic spline through a closed 2D loop and resample it.
-
-    Parameters
-    ----------
-    x, y : array_like
-        1D arrays of the polygon vertices, ordered along the loop.
-        They may include the first point again at the end; if so,
-        the duplicate endpoint is removed automatically.
-    num : int, optional
-        Number of points to sample on the smooth curve (default: 400).
-    smooth : float, optional
-        Smoothing factor passed to splprep. 0 means interpolation
-        (curve passes exactly through all input points). Larger values
-        produce smoother curves that only approximate the points.
-    k : int, optional
-        Spline degree (1 to 5). Default is cubic (3).
-
-    Returns
-    -------
-    xs, ys : ndarray
-        Sampled points along the smooth periodic curve, length `num`.
-    tck : tuple
-        Spline representation as returned by splprep (useful if you
-        want to evaluate derivatives, etc.).
-    """
-    x = np.asarray(x, dtype=float)
-    y = np.asarray(y, dtype=float)
-
-    if x.shape != y.shape:
-        raise ValueError("x and y must have the same shape")
-
-    if x.ndim != 1:
-        raise ValueError("x and y must be one-dimensional")
-
-    if x.size < k + 1:
-        raise ValueError(f"Need at least k+1={k+1} points for a degree-{k} spline")
-
-    # If the loop already repeats the first point at the end, drop the duplicate.
-    if np.allclose([x[0], y[0]], [x[-1], y[-1]]):
-        x = x[:-1]
-        y = y[:-1]
-
-    # Build a periodic parametric spline (x(u), y(u)), u in [0, 1]
-    tck, u = splprep([x, y], s=smooth, per=True, k=k)
-
-    # Sample along the spline
-    u_new = np.linspace(0.0, 1.0, num)
-    xs, ys = splev(u_new, tck)
-
-    return np.asarray(xs), np.asarray(ys), tck
